feat: reuse common logic of loading user from backend
