+++ src/views/GeneralView/components/SignIn.jsx
@@ -5,63 +5,22 @@ import {
     Col,
     Form,
     Input,
-    message,
     Row,
     Spin,
     Typography,
   } from "antd";
-  import React, { Fragment, useState } from "react";
+  import React, { Fragment } from "react";
   import { Link } from "react-router-dom";
-  import { useNavigate } from "react-router-dom";
-  import { useAuthContext } from "../../../context/AuthContext";
-//   import useScreenSize from "../../hooks/useScreenSize";
-  import { setToken } from "../../../helpers";
+import useLoadUser from "../../../api/useLoadUser";
   
   const SignIn = () => {
     const { isDesktopView } = true;
-    const navigate = useNavigate();
-  
-    const { setUser } = useAuthContext();
-  
-    const [isLoading, setIsLoading] = useState(false);
-  
-    const [error, setError] = useState("");
-  
-    const onFinish = async (values) => {
-      setIsLoading(true);
-      try {
-        const value = {
-          identifier: values.email,
-          password: values.password,
-        };
-        const response = await fetch(`${process.env.REACT_APP_BACKEND_URL}/auth/local`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(value),
-        });
-  
-        const data = await response.json();
-        if (data?.error) {
-          throw data?.error;
-        } else {
-          // set the token
-          setToken(data.jwt);
-  
-          // set the user
-          setUser(data.user);
-  
-          message.success(`Welcome back ${data.user.username}!`);
-          navigate('/', {replace:true})
-        }
-      } catch (error) {
-        console.error(error);
-        setError(error?.message ?? "Something went wrong!");
-      } finally {
-        setIsLoading(false);
-      }
-    };
+    const {
+      error,
+      isLoading,
+      setError,
+      onLoad: onFinish
+    } = useLoadUser(false);
   
     return (
       <Fragment>
